backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.6.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
beautifulsoup4==4.12.3
openai==1.50.0
anthropic==0.40.0
